requirements.txt
click
pyyaml
psutil
requests